% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myConstr.R
\name{myConstr}
\alias{myConstr}
\title{Constructor function}
\usage{
myConstr(x, y, alpha)
}
\arguments{
\item{alpha}{}
}
\value{
a list containing a data frame of the vectors, alpha, confidence interval, and the p-value
}
\description{
Takes two vectors and a alpha value, and returs a list containing a data frame of the vectors, alpha, confidence interval, and the p-value from a t-test. Assumes that the variances of the two vectors are equal.
}
\examples{
x <-rnorm(30,5,2); y<- rnorm(30, 3,2); alpha <- 0.05; myConstr(x = x, y = y, alpha = 0.05)

}
