% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotRttest.R
\name{plot.Rttest}
\alias{plot.Rttest}
\title{Plot method for the Constructor function of class Rttest}
\usage{
\method{plot}{Rttest}(x, ...)
}
\arguments{
\item{x}{data set containing population variables x and y}

\item{...}{extra parameters}
}
\value{
a ggplot with side by side boxplots of the two population samples
}
\description{
Plot method for the Constructor function of class Rttest
}
\examples{
\dontrun{ obj <- myConstr(x = x, y = y, alpha = 0.05); plot(obj)}
}
