% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.Rttest.R
\name{print.Rttest}
\alias{print.Rttest}
\title{Print method  for the constructor function with the class Rttest}
\usage{
\method{print}{Rttest}(x)
}
\arguments{
\item{x}{vector of what we are trying to get from the list}
}
\value{
the data frame x and y
}
\description{
Print method  for the constructor function with the class Rttest
}
\examples{


}
