% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myConstr.R
\name{myConstr}
\alias{myConstr}
\title{Constructor function

Takes two vectors and a alpha value, and returs a list containing a data frame of the vectors, alpha, confidence interval, and the p-value from a t-test. Assumes that the variances of the two vectors are equal.}
\usage{
myConstr(x, y, alpha)
}
\arguments{
\item{x}{vector describing population x}

\item{y}{vector describing population y}

\item{alpha}{alpha value for the t-test}
}
\value{
a list containing a data frame of the vectors, alpha, confidence interval, and the p-value
}
\description{
Constructor function

Takes two vectors and a alpha value, and returs a list containing a data frame of the vectors, alpha, confidence interval, and the p-value from a t-test. Assumes that the variances of the two vectors are equal.
}
\examples{
\dontrun{ obj <- myConstr(x = x, y = y, alpha = 0.05)}
}
