% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printRttest.R
\name{print.Rttest}
\alias{print.Rttest}
\title{Print method  for the constructor function of class Rttest}
\usage{
\method{print}{Rttest}(x, ...)
}
\arguments{
\item{x}{the data frame from the list in the constructor function}

\item{...}{extra parameters}
}
\value{
the data frame x and y
}
\description{
Print method  for the constructor function of class Rttest
}
\examples{
\dontrun{ obj <- myConstr(x = x, y = y, alpha = 0.05); print(obj)}
}
